chore: refactor all components to split up domains
