+++ website/components/Blog/index.ts
@@ -1,0 +1,1 @@
+export { default as FormattedDate } from './FormattedDate'
