+++ website/components/FinalForm/FFInput.tsx
@@ -1,5 +1,5 @@
 import { useField } from 'react-final-form'
-import Input from '../Input'
+import { Input } from '@/components/Input'
 
 const FFInput: React.FC<React.ComponentProps<typeof Input>> = (props = {}) => {
   const { input, meta } = useField(props.name)
