+++ website/components/Input/index.ts
@@ -1,0 +1,3 @@
+export { default as Button } from './Button'
+export { default as Input } from './Input'
+export { default as Switch } from './Switch'
