+++ website/components/Layout/index.ts
@@ -1,0 +1,3 @@
+export { default as Header } from './Header'
+export { default as Layout } from './Layout'
+export { default as ThemeProvider } from './ThemeProvider'
