+++ website/components/Transactions/index.ts
@@ -1,0 +1,1 @@
+export { default as TransactionCard } from './TransactionCard'
