+++ website/pages/dashboard.tsx
@@ -2,8 +2,8 @@ import { Transaction } from '@/lib/types'
 import Router from 'next/router'
 import styled from 'styled-components'
 import useSWR from 'swr'
-import Layout from '@/components/layout'
-import TransactionCard from '@/components/TransactionCard'
+import { Layout } from '@/components/Layout'
+import { TransactionCard } from '@/components/Transactions'
 import useUser from '@/hooks/useUser'
 import { linkToLogin } from '@/lib/links'
